Update ogf.h - 64b return type for ogf_find_frame and _ogf_find_frame
Fixes support for files > 2GB

+++ include/ogf.h
@@ -16,7 +16,7 @@
 
 int get_ogf_metadata(PerlIO *infile, char *file, HV *info, HV *tags);
 int ogf_find_frame_return_info(PerlIO *infile, char *file, int offset, HV *info);
-int ogf_find_frame(PerlIO *infile, char *file, int offset);
+off_t ogf_find_frame(PerlIO *infile, char *file, int offset);
 
 static int _ogf_parse(PerlIO *infile, char *file, HV *info, HV *tags, uint8_t seeking);
-static int _ogf_find_frame(PerlIO *infile, char *file, int offset, HV *info, HV *tags);
+static off_t _ogf_find_frame(PerlIO *infile, char *file, int offset, HV *info, HV *tags);
